ds-section3-sprint3-n333/Basic_Twit_Application/basic_twit_app/requirements.txt
alembic==1.4.3
bleach==3.2.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
dnspython==2.0.0
docutils==0.16
email-validator==1.1.1
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==21.4.0
Mako==1.1.3
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
pathtools==0.1.2
pkginfo==1.5.0.1
psycopg2==2.8.6
Pygments==2.7.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
readme-renderer==26.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
SQLAlchemy==1.3.20
tqdm==4.50.1
tweepy==3.9.0
twine==3.2.0
urllib3==1.25.10
watchdog==0.10.3
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3